story-enhancer-ai/requirements.txt
# ============================================
# STORY ENHANCER AI - REQUIREMENTS
# Etapa 4: Orquestación con LangGraph
# ============================================

# === CORE DEPENDENCIES ===
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0

# === RAG & VECTOR STORES ===
langchain==0.1.10
langchain-community==0.0.25
langchain-openai==0.0.8
chromadb==0.4.22
sentence-transformers==2.5.1

# === DOCUMENT PROCESSING ===
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
unstructured==0.12.4
python-magic==0.4.27

# === AGENTS & ORCHESTRATION ===
crewai==0.28.8
crewai-tools==0.1.6
langgraph==0.0.40
langsmith==0.1.0

# === UTILITIES ===
requests==2.31.0
numpy==1.26.3
pandas==2.2.0

# === OPTIONAL: LOCAL LLM SUPPORT ===
# ollama==0.1.7
# transformers==4.37.2
# torch==2.2.0

# ============================================
# STORY ENHANCER AI - REQUIREMENTS
# Etapa 4: Orquestación con LangGraph
# ============================================

# === CORE DEPENDENCIES ===
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0

# === RAG & VECTOR STORES ===
langchain==0.1.10
langchain-community==0.0.25
langchain-openai==0.0.8
chromadb==0.4.22
sentence-transformers==2.5.1

# === DOCUMENT PROCESSING ===
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
unstructured==0.12.4
python-magic==0.4.27

# === AGENTS & ORCHESTRATION ===
crewai==0.28.8
crewai-tools==0.1.6
langgraph==0.0.40
langsmith==0.1.0

# === UTILITIES ===
requests==2.31.0
numpy==1.26.3
pandas==2.2.0

# === OPTIONAL: LOCAL LLM SUPPORT ===
# ollama==0.1.7
# transformers==4.37.2
# torch==2.2.0

# Core AI/ML Libraries
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.8
langgraph==0.0.55
crewai==0.28.8
crewai-tools==0.1.6

# Vector Databases
chromadb==0.4.24
pinecone-client==3.2.2

# Embeddings & LLM
openai==1.35.3
sentence-transformers==2.7.0
transformers==4.40.2
torch==2.3.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.9
unstructured==0.13.7
markdown==3.6

# Frontend
streamlit==1.34.0
gradio==4.26.0
plotly==5.20.0
pandas==2.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
aiofiles==23.2.1
fastapi==0.111.0
uvicorn==0.29.0

# Development & Testing
pytest==8.2.1
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# System
psutil==5.9.8
tenacity==8.3.0
tqdm==4.66.4
